+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add `priority-classes` as a new app in v34.x.x releases.
+- Add `priority-classes` as a new app in v34 release.
 
 ### Fixed